Fix broken EdgeHTML issue tracker link
This fixes a broken link by replacing it with the Wayback Machine
version. This is similar to the fix in #257 but note that the bug
trackers are different: the previous fix was for a link to an Internet
Explorer bug tracker, while this fix is for a link to an EdgeHTML bug
tracker (which apparently also accepted bug reports for IE). Presumably
Microsoft discontinued this bug tracker when Edge switched to Chromium
for its browser engine.